Fix changing forum of a topic in Backend
Changing the forum of a topic which has posts in Backend failed (at
least on PostgreSQL) as due to the `uid` column being set as
`foreign_sortby` the TYPO3 RelationHandler tried to update the `uid` of
the post (!!). In order to circumvent this until RelationHandler is
fixed do not sort posts of a topic in backend.