Add notifications for forum subscriptions
It was possible to subscribe to a forum like to a topic, but
notifications for new topics in a forum were never sent as this was not
implemented. This patch is based on code by @AMartinNo1 and fixes
https://github.com/mittwald/typo3_forum/issues/184
implements the missing notifications to forum subscriptions.

Notifications are only sent once to a user even if the user subscribed
to a forum and its subforum and a topic is created in the subforum.

Translations for the new labels `Mail_Subscribe_NewPost_Subject` and
`Mail_Subscribe_NewPost_Body` have been added to English and German
only.