[BUGFIX] Send link to forum instead of topic in new topic notification
When the 'onTopicCreated' signal is sent the new topic has not been
persisted yet. This leads to current notifications for new topics being
sent with an invalid link (as it does not contain a topic uid)

Changing this would change the signal for other extensions using the
signal. Instead of doing so this patch makes the notification for a new
topic behave like one for a new post: Send a link to the parent object
(here the forum) and the name of the object (here the topic) instead.